+++ src/main/java/com/koddy/server/member/application/usecase/query/LookAroundMembersByConditionQuery.kt
@@ -3,6 +3,7 @@ package com.koddy.server.member.application.usecase.query
 import com.koddy.server.member.domain.model.Language
 import com.koddy.server.member.domain.model.Nationality
 import com.koddy.server.member.domain.repository.query.spec.SearchMenteeCondition
+import com.koddy.server.member.domain.repository.query.spec.SearchMentorCondition
 
 data class LookAroundMenteesByConditionQuery(
     val nationalities: List<Nationality>,
@@ -14,5 +15,17 @@ data class LookAroundMenteesByConditionQuery(
             return SearchMenteeCondition.basic()
         }
         return SearchMenteeCondition.of(nationalities, languages)
+    }
+}
+
+data class LookAroundMentorsByConditionQuery(
+    val languages: List<Language.Category>,
+    val page: Int,
+) {
+    fun toCondition(): SearchMentorCondition {
+        if (languages.isEmpty()) {
+            return SearchMentorCondition.basic()
+        }
+        return SearchMentorCondition.of(languages)
     }
 }
